requirements.txt
flask
click
Flask-PyMongo
python-dotenv
